refactor: batch module 추출
