+++ batch-module/src/main/java/org/c4marathon/assignment/AccountScheduler.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.scheduler;
+package org.c4marathon.assignment;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
